refactor: extract request user helper
